+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 * **build**: fixed error when the CosmWasm Optimizer image didn't exist in the local environment ([#155](https://github.com/archway-network/archway-cli/pull/155))
 * **build**: fixed error when a build container is already running in the background ([#157](https://github.com/archway-network/archway-cli/pull/157))
+* **store**: fixed error when validating the stored wasm file on-chain ([#158](https://github.com/archway-network/archway-cli/pull/158))
 
 ### Breaking Changes